Update homework1.py
Code asks the user to input total number of seconds that has past since midnight.  After getting the input from user, the code first uses the //  operator (// 3600) to output only the whole number of 1 for the number of hours.  Then  to compute the number of minutes, the code again refers back to the total number of seconds that were input  and finds the remainder left when total seconds are divided by 3600 using the modulus operand (%) after which that number is divided by 60 again only getting the whole number (//).  Finally for the number of seconds, the total number of seconds can be found by 1st getting the remainder of seconds using % 3600 (200 seconds left), and eliminating the minutes so that only seconds remain using % 60 since there's 60 seconds in a minute.